backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator>=2.1.0
python-multipart==0.0.20
anthropic>=0.40.0
google-api-python-client>=2.150.0
google-auth>=2.36.0
pdfplumber>=0.11.0
pytesseract>=0.3.13
python-docx>=1.1.0
beautifulsoup4>=4.12.0
chardet>=5.2.0
weasyprint>=62.0
apscheduler>=3.10.0
httpx>=0.28.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
